backend/requirements.txt
# Django and REST Framework
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
drf-spectacular==0.27.0

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0
# Core Django
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
drf-spectacular==0.27.0
django-filter==23.5

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Cache & Tasks
django-redis==5.4.0
redis==5.0.1
celery==5.3.6

# Deployment
gunicorn==21.2.0
whitenoise==6.6.0

# Utilities
python-decouple==3.8
Pillow==10.2.0
# Celery and Redis
celery==5.3.6
redis==5.0.1

# Server and Static Files
gunicorn==21.2.0
whitenoise==6.6.0

# Other utilities
python-decouple==3.8
Pillow==10.2.0
